Add BaseEncoder Class
